refactor(dev-infra): move the release notes into its own directory under release
Move the release notes tooling under its own directory under release rather than
within publish, in preparation to have a release note generation command for ad-hoc
release note generation.
