+++ dev-infra/release/notes/BUILD.bazel
@@ -1,20 +1,17 @@
 load("@npm//@bazel/typescript:index.bzl", "ts_library")
 
 ts_library(
-    name = "publish",
+    name = "notes",
     srcs = glob([
         "**/*.ts",
     ]),
-    module_name = "@angular/dev-infra-private/release/publish",
+    module_name = "@angular/dev-infra-private/release/notes",
     visibility = ["//dev-infra:__subpackages__"],
     deps = [
         "//dev-infra/commit-message",
-        "//dev-infra/pr/merge",
         "//dev-infra/release/config",
-        "//dev-infra/release/notes",
         "//dev-infra/release/versioning",
         "//dev-infra/utils",
-        "@npm//@octokit/rest",
         "@npm//@types/ejs",
         "@npm//@types/inquirer",
         "@npm//@types/node",
@@ -22,8 +19,6 @@ ts_library(
         "@npm//@types/yargs",
         "@npm//ejs",
         "@npm//inquirer",
-        "@npm//ora",
         "@npm//semver",
-        "@npm//typed-graphqlify",
     ],
 )
