+++ dev-infra/release/publish/constants.ts
@@ -6,11 +6,11 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-/** Project-relative path for the changelog file. */
-export const changelogPath = 'CHANGELOG.md';
-
 /** Project-relative path for the "package.json" file. */
 export const packageJsonPath = 'package.json';
+
+/** Project-relative path for the changelog file. */
+export const changelogPath = 'CHANGELOG.md';
 
 /** Default interval in milliseconds to check whether a pull request has been merged. */
 export const waitForPullRequestInterval = 10000;
